Format project (#13)
* Add UsersController Unit Tests

* Nicer error message from GetValue helper

GetValueFromObjectResult previously allowed the exception to fail the
test, but it's more usable to catch the exception and Assert.Fail with
the message.

Also added .editorconfig to Solution Items so formatting is consistent.

* Fix formatting

Visual Studio totally butchered the tabbing on these files and I'm not
sure why. I formatted them all to the editorconfig, and will make a
followup PR to unify the rest of the formatting.

* Change GetValueFromObjectResult line setup

Syntax highlighting didn't work in VSCode or GitHub for some reason with
this setup.

* Change test fixture class name

* Format entire project

The project had inconsistent tabs/spaces and line endings, this should
unify everything to the editorconfig settings.