Removed integration project, replaced unit tests
As I continued to improve Integration tests, I realized that they are
going to be essentially as effective, and (when using an in memory db)
barely slower than unit tests. Might as well save ourselves tons of
effort and have minimal tests that basically just test the API as if it
were a real client.